+++ src/domain/Post/useCases/usePostList.ts
@@ -1,64 +1,5 @@
-import {useEffect, useState} from 'react';
-
-import {Post, postService} from '@domain';
+import {Post, postService, usePaginatedList} from '@domain';
 
 export function usePostList() {
-  const [postList, setPostList] = useState<Post[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<boolean | null>(null);
-  const [page, setPage] = useState(1);
-  const [hasNextPage, setHasNextPage] = useState(true);
-
-  async function fetchInitialData() {
-    try {
-      setError(null);
-      setLoading(true);
-      const {data, meta} = await postService.getList(1);
-      setPostList(data);
-      if (meta.hasNextPage) {
-        setPage(2);
-        setHasNextPage(true);
-      } else {
-        setHasNextPage(false);
-      }
-    } catch (er) {
-      console.log('ERROR', error);
-      setError(true);
-    } finally {
-      setLoading(false);
-    }
-  }
-
-  async function fetchNextPage() {
-    if (loading || !hasNextPage) {
-      return;
-    }
-    try {
-      setLoading(true);
-      const {data, meta} = await postService.getList(page);
-      setPostList(prev => [...prev, ...data]);
-      if (meta.hasNextPage) {
-        setPage(prev => prev + 1);
-      } else {
-        setHasNextPage(false);
-      }
-    } catch (er) {
-      setError(true);
-    } finally {
-      setLoading(false);
-    }
-  }
-
-  useEffect(() => {
-    fetchInitialData();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  return {
-    postList,
-    error,
-    loading,
-    refresh: fetchInitialData,
-    fetchNextPage,
-  };
+  return usePaginatedList<Post>(postService.getList);
 }
