+++ src/domain/PostComment/postCommentService.ts
@@ -22,6 +22,6 @@ async function getList(
   };
 }
 
-export const postService = {
+export const postCommentService = {
   getList,
 };
