+++ src/domain/PostComment/useCases/usePostCommentList.ts
@@ -1,0 +1,10 @@
+import {usePaginatedList} from 'src/domain/hooks/usePaginatedList';
+
+import {postCommentService} from '../postCommentService';
+
+export function usePostCommentList(postId: number) {
+  function getList(page: number) {
+    return postCommentService.getList(postId, page);
+  }
+  return usePaginatedList(getList);
+}
