+++ src/domain/index.ts
@@ -1,3 +1,5 @@
 export * from './Post/postService';
 export * from './Post/postTypes';
 export * from './Post/useCases/usePostList';
+
+export * from './hooks/usePaginatedList';
