+++ src/screens/app/HomeScreen/HomeScreen.tsx
@@ -15,7 +15,13 @@ import {AppTabScreenProps} from '@routes';
 import {HomeHeader, HomeEmpty} from '@screens';
 
 export function HomeScreen({}: AppTabScreenProps<'HomeScreen'>) {
-  const {postList, error, loading, refresh, fetchNextPage} = usePostList();
+  const {
+    list: postList,
+    error,
+    loading,
+    refresh,
+    fetchNextPage,
+  } = usePostList();
 
   const flatListRef = useRef<FlatList<Post>>(null);
   useScrollToTop(flatListRef);
